kbuild: introduce scripts install helper
It is a hard time for out-tree package build to get the right
script file for building external modules. For in-tree build,
it is possible to invoke "install-extmod-build" script to
simplify the build process. As this script use many variables
from kbuild, it is hard to execute it directly in the out-tree
build. Export this helper script should simplify the package
build.

Introduce "script_install" target to simplify the package build.

Signed-off-by: Inochi Amaoto <inochiama@gmail.com>